Basic Navigation
Set up basic navigation between
1. Home Screen
2. Service Details Screen
3. Booking Summary Screen
