+++ app/build.gradle
@@ -55,16 +55,18 @@ dependencies {
     implementation "androidx.compose.ui:ui:$compose_version"
     implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
     implementation "androidx.compose.material3:material3:1.0.1"
-    implementation "com.google.accompanist:accompanist-systemuicontroller:0.28.0"
-    implementation 'com.google.firebase:firebase-database-ktx:20.0.4'
-    implementation 'com.google.code.gson:gson:2.10'
+    implementation "androidx.navigation:navigation-compose:2.5.3"
 
-    testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.4'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
     androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
     debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
     debugImplementation "androidx.compose.ui:ui-test-manifest:$compose_version"
 
+    implementation "com.google.accompanist:accompanist-systemuicontroller:0.28.0"
+    implementation 'com.google.firebase:firebase-database-ktx:20.0.4'
+    implementation 'com.google.code.gson:gson:2.10'
     implementation platform('com.google.firebase:firebase-bom:31.1.1')
+
+    testImplementation 'junit:junit:4.13.2'
 }
