+++ app/src/main/java/com/felix/felix/ui/components/ServiceCard.kt
@@ -29,7 +29,8 @@ fun ServiceCard(
     price: String,
     caption: String = "",
     buttonVisible : Boolean = true,
-    shape : Shape = CardDefaults.elevatedShape
+    shape : Shape = CardDefaults.elevatedShape,
+    onClick : () -> Unit = {}
 ) {
     ElevatedCard(
         shape = shape,
@@ -37,7 +38,7 @@ fun ServiceCard(
             .fillMaxWidth()
             .height(180.dp)
             .padding(horizontal = 25.dp),
-        onClick = {/*TODO: On offer card click*/}
+        onClick = onClick
     ) {
         Row(
             Modifier.fillMaxHeight()
@@ -58,7 +59,7 @@ fun ServiceCard(
                         verticalArrangement = Arrangement.Bottom
                     ) {
                         OutlinedButton(
-                            onClick = { /*TODO*/ },
+                            onClick = { onClick() },
                             shape = MaterialTheme.shapes.small,
                             modifier = Modifier.size(width = 120.dp, height = 40.dp),
                             contentPadding = PaddingValues(horizontal = 8.dp),
